requirements.txt
CTkMessagebox==2.7
customtkinter==5.2.2
matplotlib==3.10.3
numpy==2.3.1
open3d==0.19.0
opencv_python==4.11.0.86
Pillow==11.2.1
scikit_learn==1.7.0
tqdm==4.67.1
trimesh==4.6.8
